# docs/swagger.yaml
basePath: /
definitions:
  chat.Dialog:
    properties:
      castingID:
        type: string
      createdAt:
        type: string
      id:
        type: string
      imageURL:
        type: string
      isGroup:
        type: boolean
      lastMessage:
        $ref: '#/definitions/chat.Message'
      lastMessageID:
        type: string
      participants:
        items:
          $ref: '#/definitions/chat.DialogParticipant'
        type: array
      title:
        type: string
      updatedAt:
        type: string
    type: object
  chat.DialogParticipant:
    properties:
      dialogID:
        type: string
      id:
        type: string
      isMuted:
        type: boolean
      joinedAt:
        type: string
      lastSeenAt:
        type: string
      leftAt:
        type: string
      role:
        description: member, admin, owner
        type: string
      typingUntil:
        type: string
      userID:
        type: string
    type: object
  chat.Message:
    properties:
      attachmentName:
        type: string
      attachmentURL:
        type: string
      attachments:
        items:
          $ref: '#/definitions/chat.MessageAttachment'
        type: array
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      dialogID:
        type: string
      forwardFrom:
        $ref: '#/definitions/chat.Message'
      forwardFromID:
        type: string
      id:
        type: string
      reactions:
        items:
          $ref: '#/definitions/chat.MessageReaction'
        type: array
      readReceipts:
        items:
          $ref: '#/definitions/chat.MessageReadReceipt'
        type: array
      replyTo:
        $ref: '#/definitions/chat.Message'
      replyToID:
        type: string
      senderID:
        type: string
      status:
        description: sent, delivered, read, deleted
        type: string
      type:
        description: text, image, video, file, system
        type: string
    type: object
  chat.MessageAttachment:
    properties:
      createdAt:
        type: string
      fileName:
        type: string
      fileType:
        description: image, video, file
        type: string
      id:
        type: string
      messageID:
        type: string
      mimeType:
        type: string
      size:
        type: integer
      uploaderID:
        type: string
      url:
        type: string
    type: object
  chat.MessageReaction:
    properties:
      createdAt:
        type: string
      emoji:
        type: string
      id:
        type: string
      messageID:
        type: string
      userID:
        type: string
    type: object
  chat.MessageReadReceipt:
    properties:
      id:
        type: string
      messageID:
        type: string
      readAt:
        type: string
      userID:
        type: string
    type: object
  dto.CreateDialogInput:
    description: Структура запроса для создания личного или группового чата creator_id
      должен быть одним из участников participant_ids - список user_id, включая creator
      is_group = true означает групповой чат, иначе личный title используется только
      для группового чата
    properties:
      creator_id:
        type: string
      is_group:
        type: boolean
      participant_ids:
        items:
          type: string
        type: array
      title:
        description: nullable — только для групп
        type: string
    required:
    - creator_id
    - participant_ids
    type: object
  dto.ModelAnalytics:
    properties:
      income:
        type: number
      rating:
        type: number
      responses:
        type: integer
      views:
        type: integer
    type: object
  dto.SendMessageInput:
    description: Используется при отправке сообщения, включая вложения и цитирование
    properties:
      attachment_ids:
        items:
          type: string
        type: array
      content:
        type: string
      dialog_id:
        type: string
      forward_from:
        type: string
      reply_to_id:
        type: string
      sender_id:
        type: string
    required:
    - dialog_id
    - sender_id
    type: object
...
  handlers.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.RefreshInput:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  handlers.RegisterInput:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      role:
        type: string
    required:
    - email
    - password
    - role
    type: object
...
  handlers.toggleReactionInput:
    properties:
      emoji:
        type: string
      message_id:
        type: string
      user_id:
        type: string
    required:
    - emoji
    - message_id
    - user_id
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        example: invalid input
        type: string
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      email:
        description: уникальный индекс
        type: string
      id:
        description: UUID primary key
        type: string
      isVerified:
        type: boolean
      passwordHash:
        type: string
      resetToken:
        type: string
      resetTokenExp:
        type: string
      role:
        description: model | employer | admin
        type: string
      status:
        description: pending | active | suspended | banned
        type: string
      subscription:
        description: free | premium | pro
        type: string
      updatedAt:
        type: string
      verificationToken:
        type: string
    type: object
host: localhost:4000
info:
  contact:
    email: support@company.com
    name: Модельное агенство
    url: http://contact.company.com
  description: API для управления модельной компанией (документация Swagger).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://your-terms.com/terms/
  title: mwork API
  version: "1.0"
paths:
  /analytics/model:
    get:
      consumes:
      - application/json
      description: Возвращает статистику по просмотрам, рейтингу, доходу и откликам
        модели. UserID берётся из JWT.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ModelAnalytics'
        "400":
          description: Некорректный userID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизован или отсутствует токен
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Профиль модели не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка при получении аналитики
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить аналитику модели
      tags:
      - analytics
  /auth/login:
    post:
      consumes:
      - application/json
      description: Возвращает access и refresh токены при успешной аутентификации
      parameters:
      - description: Email и пароль
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: access_token, refresh_token, user
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный формат ввода
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неверные учётные данные
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Вход пользователя
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Удаляет refresh токен (разлогинивает пользователя)
      parameters:
      - description: Refresh токен для удаления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshInput'
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный формат ввода
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Ошибка при удалении токена
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Выход из системы
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Получить новый access-токен по refresh-токену
      parameters:
      - description: Refresh токен
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshInput'
      produces:
      - application/json
      responses:
        "200":
          description: access_token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный формат ввода
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Недействительный токен
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Обновление access-токена
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя по email, паролю и роли
      parameters:
      - description: Данные регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Неверный формат ввода или ошибка регистрации
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Регистрация нового пользователя
      tags:
      - auth
  /chat/dialogs:
    post:
      consumes:
      - application/json
      parameters:
      - description: Информация о диалоге
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDialogInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/chat.Dialog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Создать новый диалог (личный или групповой)
      tags:
      - Chat
  /chat/dialogs/{id}/files:
    get:
      parameters:
      - description: ID диалога
        in: path
        name: id
        required: true
        type: string
      - description: Фильтр по типу (image, video, file)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/chat.MessageAttachment'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить вложения в диалоге
      tags:
      - Chat
  /chat/dialogs/{id}/leave:
    post:
      parameters:
      - description: ID диалога
        in: path
        name: id
        required: true
        type: string
      - description: ID пользователя
        in: query
        name: user_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Выйти из диалога
      tags:
      - Chat
  /chat/dialogs/{id}/messages:
    get:
      parameters:
      - description: ID диалога
        in: path
        name: id
        required: true
        type: string
      - description: ID пользователя
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/chat.Message'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить сообщения из диалога
      tags:
      - Chat
  /chat/dialogs/{id}/read:
    post:
      parameters:
      - description: ID диалога
        in: path
        name: id
        required: true
        type: string
      - description: ID пользователя
        in: query
        name: user_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Отметить все сообщения как прочитанные
      tags:
      - Chat
  /chat/dialogs/{id}/unread:
    get:
      parameters:
      - description: ID диалога
        in: path
        name: id
        required: true
        type: string
      - description: ID пользователя
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить количество непрочитанных сообщений в диалоге
      tags:
      - Chat
  /chat/messages/send:
    post:
      consumes:
      - application/json
      parameters:
      - description: Тело сообщения
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.SendMessageInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/chat.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Отправить сообщение в диалог
      tags:
      - Chat
  /chat/reactions/toggle:
    post:
      consumes:
      - application/json
      parameters:
      - description: Реакция
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.toggleReactionInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Поставить или убрать реакцию на сообщение
      tags:
      - Chat
...
swagger: "2.0"